[UPDATE] Location Float -> Double

+++ Framework/DFINERY-SDK/Model/Req/Location/Location.swift
@@ -1,6 +1,6 @@
 import Foundation
 
 struct UserLocation{
-    let lat : Float
-    let lng : Float
+    let lat : Double
+    let lng : Double
 }
